operators: propagate dtype into AIERuntimeArgSpec
strided_copy/repeat/gemm build AIERuntimeArgSpec without a dtype kwarg, so
every returned spec silently reports the dataclass default (bfloat16)
regardless of the operator's own dtype field(s). The two real consumers
(test_utils.py's XRTTensor(spec.shape, dtype=spec.dtype), sequence.py's
calculate_buffer_layout) size buffers off spec.dtype, so a non-default-dtype
caller under- or over-allocates by the itemsize ratio -- GEMM(dtype_in='i8',
dtype_out='i32') declared a 2-byte C buffer for an actual 4-byte int32
output. Latent: every in-tree caller uses the default today. Mirrors
dequant/op.py's existing correct pattern.

+++ iron/operators/repeat/op.py
@@ -56,8 +56,10 @@ class Repeat(MLIROperator):
 
     def get_arg_spec(self):
         return [
-            AIERuntimeArgSpec("in", (self.rows, self.cols)),
-            AIERuntimeArgSpec("out", (self.rows * self.repeat, self.cols)),
+            AIERuntimeArgSpec("in", (self.rows, self.cols), dtype=self.dtype),
+            AIERuntimeArgSpec(
+                "out", (self.rows * self.repeat, self.cols), dtype=self.dtype
+            ),
         ]
 
     def reference(self, x):
